+++ tests/expected/func_calls_variadic_parameters.gd
@@ -1,0 +1,11 @@
+func only_variadic(...args):
+	pass
+
+
+func call_with_many_arguments(
+	first_argument: String,
+	second_argument: String,
+	third_argument: String,
+	...args: Array,
+):
+	pass