+++ tests/input/setget.gd
@@ -2,7 +2,7 @@
 # Tests both setter and getter syntaxes.
 var  a  :  float  =  10.0   :
 	set(  value   )  :    a = value
-	get   :    return a
+	get()   :    return a
 var second: int = 10
 
 var b: float = 10.0: