+++ CHANGELOG.md
@@ -11,6 +11,9 @@ This file documents the changes made to the formatter with each release.
 ### Changed
 
 - Removed space between lambda function name and parameter list
+- Force @export and @onready annotations to stay on the same line as a variable but keep other annotations separate
+- Stop trying to format any code containing parse errors. Until now we tried to still format definitions around the code with errors, but this can lead to cases where the formatter produces invalid code
+- Add line wrapping for annotation arguments (#330)
 
 ### Fixed
 
@@ -18,6 +21,11 @@ This file documents the changes made to the formatter with each release.
 - fixed certain export annotations being moved out of their respective groups (#308)
 - Preserve up to one blank line used to group elements in "containers" like enums
 - Fixed losing blank line between statements in a body if the previous statement has an inline comment (#320, thanks @Buitragox for the fix)
+- Fixed various edge cases with ignored directories: paths are now normalized before comparison
+- Fixed special get syntax with parentheses having an extra space (#314)
+- Output warnings to stderr when using reorder and safe mode together (#332)
+- Removed the space between `...` and variadic parameter names (#331)
+- Fixed incorrect indentation for match patterns using `and` or `or` (#328)
 
 ## Release 0.24.0 (2026-07-25)